+++ src/header/LED.h
@@ -1,8 +1,9 @@
 #include "FastLED.h"
 #include "release.h"
+#include "stdlib.h"
 
-constexpr int num_leds = rowcount * colcount;
-constexpr int led_pin = 6;
+constexpr uint16_t num_leds = rowcount * colcount;
+constexpr uint8_t led_pin = 6;
 
 extern CRGB leds[num_leds];
