+++ src/header/matrix.h
@@ -14,8 +14,8 @@ public:
 
     // Methodds
 
-    CRGB get_LED(int, int) const;
-    void set_LED(CRGB, int, int);
+    CRGB get_LED(uint8_t, uint8_t) const;
+    void set_LED(CRGB, uint8_t, uint8_t);
     void matrix_to_LEDArray(CRGB *) const;
     void shift_Left();
     void replace_last_col(std::vector<CRGB>);
